backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.7
pydantic-settings>=2.2
google-genai>=0.8
sqlalchemy>=2.0
psycopg2-binary>=2.9
pgvector>=0.3
alembic>=1.13
